Reorganized ProfilePage functions into separate files
