+++ app/frontend/src/pages/profile/components/ProfileHeader.tsx
@@ -1,0 +1,18 @@
+import { Paper } from "@mui/material";
+import { styled } from "@mui/material/styles";
+
+const Item = styled(Paper)(({ theme }) => ({
+  backgroundColor: theme.palette.mode === "dark" ? "#1A2027" : "#fff",
+  ...theme.typography.body2,
+  padding: theme.spacing(1),
+  textAlign: "center",
+  color: theme.palette.text.secondary
+}));
+
+export function ProfileHeader() {
+  return (
+    <>
+      <Item>Profile Section</Item>
+    </>
+  );
+}
