+++ app/frontend/src/routes/ProfileRoute.tsx
@@ -1,4 +1,4 @@
-import ProfilePage from "src/pages/ProfilePage";
+import ProfilePage from "src/pages/profile/ProfilePage";
 
 export default function ProfileRoute() {
   return (
